# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Solar Generation API",
    "version": "1.0.0",
    "description": "Implemented endpoints: application liveness, OpenAPI, user/admin and device JWT exchange, and jurisdiction-scoped province lists. Active installations can submit append-only readings. Users can retrieve individual readings within their stored jurisdiction. Users can page and filter installation reading history within current jurisdiction."
  },
  "servers": [
    {
      "url": "/api/v1.0"
    }
  ],
  "paths": {
...
    "/auth/user-tokens": {
      "post": {
        "operationId": "createUserToken",
        "summary": "Exchange user/admin credentials for a short-lived JWT",
        "security": [],
        "description": "Accept only email and password. Normalize email by trimming/lowercasing; preserve the password exactly. Claims come from the current MongoDB User: sub (public UUID), actor=user, role, readScope, optional provinceId or districtId; admins also receive installation-create, installation-deactivate, installation-delete in permissions. HS256 signatures include configured iss, aud, iat, and exp. No refresh token. Responses are not cacheable. Validly shaped attempts use shared MongoDB limits of 5 per 15 minutes by IP and normalized account email. No passwords, hashes, or internal IDs are returned.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "additionalProperties": false,
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1,
                    "writeOnly": true,
                    "description": "Must contain a non-whitespace character; surrounding spaces are significant."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/UserTokenIssued"
          },
          "400": {
            "description": "Malformed JSON or invalid request fields.",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "fields": {
                    "value": {
                      "code": "INVALID_REQUEST",
                      "message": "Provide only a valid email and a nonempty password.",
                      "details": []
                    }
                  },
                  "json": {
                    "value": {
                      "code": "INVALID_JSON",
                      "message": "Request body must be valid JSON.",
                      "details": []
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid email or password.",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              },
              "WWW-Authenticate": {
                "$ref": "#/components/headers/BearerChallenge"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "INVALID_CREDENTIALS",
                  "message": "Invalid email or password.",
                  "details": []
                }
              }
            }
          },
          "406": {
            "$ref": "#/components/responses/TokenNotAcceptable"
          },
          "413": {
            "$ref": "#/components/responses/TokenPayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/TokenUnsupportedMediaType"
          },
          "429": {
            "$ref": "#/components/responses/TokenRateLimited"
          },
          "500": {
            "$ref": "#/components/responses/TokenServerError"
          }
        }
      }
    },
...
    "/districts/{districtId}": {
      "get": {
        "operationId": "getDistrict",
        "summary": "Retrieve public district details within the current user jurisdiction",
        "description": "Require current user authentication (installation tokens return 401), shared 120/minute User read limits and UUID v4 districtId (invalid format 400). Resolve District and Province in a read snapshot; missing district or ancestry returns 404. National analysts/admins read nationally; provincial analysts within their stored province; district analysts only their assigned district; otherwise 403. Authorize before data or validators. Return only public id/provinceId/name, without related collections. No query options; supplied parameters return 400 INVALID_QUERY. Strong ETag covers current principal and public representation; private, no-cache. Matching If-None-Match returns bodyless 304 after current access/rate checks. No reliable district metadata Last-Modified exists; omit it and return 200 for If-Modified-Since alone. Errors use no-store without validators.",
        "security": [
          {
            "UserBearer": []
          }
        ],
        "parameters": [
          {
            "name": "districtId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Strong/weak matching tag, quoted tag list, or * returns bodyless 304 after authorization. Takes precedence over If-Modified-Since."
          },
          {
            "name": "If-Modified-Since",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Ignored because no reliable district metadata Last-Modified exists; alone returns 200."
          }
        ],
        "responses": {
          "200": {
            "description": "Public district details",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/ScopedETag"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateCacheControl"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/District"
                }
              }
            }
          },
          "304": {
            "description": "Authorized public district unchanged; no body or Content-Type",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/ScopedETag"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateCacheControl"
              }
            }
          },
          "400": {
            "description": "Invalid district UUID (INVALID_REQUEST) or unsupported query parameters (INVALID_QUERY).",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or installation token, or deleted/invalid user (UNAUTHORIZED).",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              },
              "WWW-Authenticate": {
                "$ref": "#/components/headers/BearerChallenge"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "District outside stored jurisdiction (FORBIDDEN).",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "FORBIDDEN",
                  "message": "The district is outside your permitted jurisdiction.",
                  "details": []
                }
              }
            }
          },
          "404": {
            "description": "Missing district or province ancestry (NOT_FOUND).",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "NOT_FOUND",
                  "message": "District not found.",
                  "details": []
                }
              }
            }
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "429": {
            "description": "Shared user-read limit exceeded (RATE_LIMIT_EXCEEDED).",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              },
              "Retry-After": {
                "$ref": "#/components/headers/LoginRetryAfter"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected persistence failure (INTERNAL_SERVER_ERROR).",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/provinces/{provinceId}/districts": {
      "get": {
        "operationId": "listProvinceDistricts",
        "summary": "List authorized districts belonging to a province",
        "description": "Require current user authentication (installation tokens return 401), shared 120/minute User read limits and UUID v4 provinceId (invalid format 400). Missing requested province returns 404. National analysts/admins list all districts in the requested province; provincial analysts within assigned province; district analysts may access their own province but receive only their assigned district. Other provinces or broken assigned district ancestry return 403. Authorize before district result queries and filter in MongoDB by requested provinceId plus assigned publicId for district scope. Read parent/list in one snapshot. Return count/items only with public id/provinceId/name, ordered by name/publicId ascending; empty authorized province returns 200. No pagination, sorting or geography query options; supplied parameters return 400. ETag covers current principal, province and complete filtered response; private, no-cache. Matching If-None-Match returns bodyless 304 after current access/rate checks. Omit Last-Modified; If-Modified-Since alone returns 200. Errors use no-store without validators/data/counts.",
        "security": [
          {
            "UserBearer": []
          }
        ],
        "parameters": [
          {
            "name": "provinceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Matching strong/weak tag, tag in a list, or * returns bodyless 304 only after authorization and read limiting."
          }
        ],
        "responses": {
          "200": {
            "description": "Complete authorized province district collection with count/items only",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateCacheControl"
              },
              "ETag": {
                "$ref": "#/components/headers/ScopedETag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistrictList"
                }
              }
            }
          },
          "304": {
            "description": "Complete authorized response unchanged; no body or Content-Type.",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateCacheControl"
              },
              "ETag": {
                "$ref": "#/components/headers/ScopedETag"
              }
            }
          },
          "400": {
            "description": "Invalid province UUID (INVALID_REQUEST) or unsupported query parameters (INVALID_QUERY).",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired token, installation token, or deleted/invalid user (UNAUTHORIZED).",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              },
              "WWW-Authenticate": {
                "$ref": "#/components/headers/BearerChallenge"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Requested province outside current stored jurisdiction or broken district assignment ancestry (FORBIDDEN).",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "FORBIDDEN",
                  "message": "The province is outside your permitted jurisdiction.",
                  "details": []
                }
              }
            }
          },
          "404": {
            "description": "Missing requested province (NOT_FOUND).",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "NOT_FOUND",
                  "message": "Province not found.",
                  "details": []
                }
              }
            }
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "429": {
            "description": "Shared user-read limit exceeded (RATE_LIMIT_EXCEEDED).",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              },
              "Retry-After": {
                "$ref": "#/components/headers/LoginRetryAfter"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected persistence failure (INTERNAL_SERVER_ERROR).",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message",
          "details"
        ],
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {}
          }
        }
      },
...
      "UserAccessToken": {
        "type": "object",
        "required": [
          "access_token",
          "token_type",
          "expires_in",
          "userId"
        ],
        "additionalProperties": false,
        "properties": {
          "access_token": {
            "$ref": "#/components/schemas/AccessToken/properties/access_token"
          },
          "token_type": {
            "$ref": "#/components/schemas/AccessToken/properties/token_type"
          },
          "expires_in": {
            "$ref": "#/components/schemas/AccessToken/properties/expires_in"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "description": "Authenticated public UUID from the stored principal; equals the JWT sub claim."
          }
        }
      },
...
      "District": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "provinceId",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
          },
          "provinceId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
          },
          "name": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "DistrictList": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "count",
          "items"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "minimum": 0,
            "description": "Number of authorized districts returned in the requested province."
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/District"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "UserBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "HS256 user JWT from /auth/user-tokens. Each protected request reloads the current User role and jurisdiction."
      },
      "InstallationBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Installation JWT from /auth/device-tokens; requires installation-write scope and current active installation ownership."
      }
    },
    "headers": {
      "NoStore": {
        "schema": {
          "type": "string",
          "enum": [
            "no-store"
          ]
        }
      },
      "BearerChallenge": {
        "schema": {
          "type": "string",
          "enum": [
            "Bearer"
          ]
        }
      },
      "LoginRetryAfter": {
        "schema": {
          "type": "integer",
          "minimum": 1
        },
        "description": "Seconds until the exceeded counter window expires."
      },
      "PrivateCacheControl": {
        "schema": {
          "type": "string",
          "enum": [
            "private, no-cache"
          ]
        }
      },
      "ScopedETag": {
        "schema": {
          "type": "string"
        },
        "description": "Strong validator of the exact scoped representation and current stored principal."
      },
...
    },
    "responses": {
      "TokenNotAcceptable": {
        "description": "Accept excludes JSON; no body",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/NoStore"
          }
        }
      },
      "TokenPayloadTooLarge": {
        "description": "Request body is too large.",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/NoStore"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "code": "PAYLOAD_TOO_LARGE",
              "message": "Request body is too large.",
              "details": []
            }
          }
        }
      },
      "TokenUnsupportedMediaType": {
        "description": "Request body must use application/json.",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/NoStore"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "code": "UNSUPPORTED_MEDIA_TYPE",
              "message": "Request body must use application/json.",
              "details": []
            }
          }
        }
      },
      "TokenRateLimited": {
        "description": "Too many login attempts. Try again later.",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/NoStore"
          },
          "Retry-After": {
            "$ref": "#/components/headers/LoginRetryAfter"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "code": "RATE_LIMIT_EXCEEDED",
              "message": "Too many login attempts. Try again later.",
              "details": []
            }
          }
        }
      },
      "TokenServerError": {
        "description": "An unexpected error occurred.",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/NoStore"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "code": "INTERNAL_SERVER_ERROR",
              "message": "An unexpected error occurred.",
              "details": []
            }
          }
        }
      },
      "NotAcceptable": {
        "description": "Accept excludes JSON; no body"
      },
      "UserTokenIssued": {
        "description": "Credentials verified; signed access token",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/NoStore"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/UserAccessToken"
            }
          }
        }
      },
...
    }
  }
}